make parser module
